+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.23.0 (2024-09-03)
+
 - added: (ADA) `isStakeTx` to `EdgeTransaction.otherParams` for staking transaction signing
 - added: (ADA) staking related methods to `EdgeCurrencyEngine.otherMethods`
 - changed: (ADA) Include `accountKey` and `privateKey` in Cardano wallet's private key data as an optimization for new wallets going forward.